+++ packages/system/src/xydrag/utils.ts
@@ -81,10 +81,12 @@ export function getEventHandlerParams<NodeType extends InternalNodeBase>({
   nodeId,
   dragItems,
   nodeLookup,
+  dragging = true,
 }: {
   nodeId?: string;
   dragItems: Map<string, NodeDragItem>;
   nodeLookup: Map<string, NodeType>;
+  dragging?: boolean;
 }): [NodeBase, NodeBase[]] {
   const nodesFromDragItems: NodeBase[] = [];
 
@@ -95,6 +97,7 @@ export function getEventHandlerParams<NodeType extends InternalNodeBase>({
       nodesFromDragItems.push({
         ...node,
         position: dragItem.position,
+        dragging,
       });
     }
   }
@@ -109,6 +112,7 @@ export function getEventHandlerParams<NodeType extends InternalNodeBase>({
     {
       ...node,
       position: dragItems.get(nodeId)?.position || node.position,
+      dragging,
     },
     nodesFromDragItems,
   ];